+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Major change: New usage, so that we can avoid having to do monkey patching [#10]
 - Show positional arguments in their own panel by default
 - Add config `GROUP_ARGUMENTS_OPTIONS` option to group with options
 - Improve handing of metavars, give option to show appended instead of in column
+- Add `COLOR_SYSTEM` option to add ability to disable colours
+- Add options to customise error message help texts
 - Add support for printing errors nicely
 
 ## Version 0.3.0 (2022-02-13)